requirements.txt
requests>=2.31.0
google-search-results>=2.4.2
SpeechRecognition>=3.10.0
pyaudio>=0.2.11
python-dotenv>=1.0.0
elevenlabs>=1.0.0
streamlit>=1.28.0
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.0.0
firebase-admin>=6.5.0
